pyproject.toml
[tool.poetry]
name = "backup-s3"
version = "0.1.0"
description = ""
authors = ["poso-douglas <douglas.poso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.36.14"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
